Change fog_provider.rb into a simple loader and move fog_provider common code to lib
