+++ provisioner/daemon/plugins/providers/fog_provider/fog_provider.rb
@@ -16,61 +16,5 @@
 # limitations under the License.
 #
 
-gem 'fog', '~> 1.21.0'
-
-require 'fog'
-require_relative 'lib/fog_provider/openstack'
+# require_relative 'lib/fog_provider/openstack'
 require_relative 'lib/fog_provider/rackspace'
-
-class FogProvider < Provider
-
-  # used by ssh validation in confirm stage
-  def set_credentials(sshauth)
-    @credentials = Hash.new
-    @credentials[:paranoid] = false
-    sshauth.each do |k, v|
-      if (k =~ /password/)
-        @credentials[:password] = v
-      elsif (k =~ /identityfile/)
-        @credentials[:keys] = [ v ]
-      end
-    end
-  end
-
-  # Open a connection to a port
-  def tcp_test_port(hostname, port)
-    tcp_socket = TCPSocket.new(hostname, port)
-    readable = IO.select([tcp_socket], nil, nil, 5)
-    if readable
-      log.debug("sshd accepting connections on #{hostname} port #{port}, banner is #{tcp_socket.gets}")
-      yield
-      true
-    else
-      false
-    end
-  rescue Errno::ETIMEDOUT
-    false
-  rescue Errno::EPERM
-    false
-  rescue Errno::ECONNREFUSED
-    sleep 2
-    false
-  rescue Errno::EHOSTUNREACH, Errno::ENETUNREACH
-    sleep 2
-    false
-  rescue Errno::ENETUNREACH
-    sleep 2
-    false
-  ensure
-    tcp_socket && tcp_socket.close
-  end
-
-  def wait_for_sshd
-    log.debug 'Waiting for sshd'
-    until tcp_test_port(bootstrap_ip, 22) {
-      sleep @initial_sleep_delay ||= 10
-      log.info "Server #{server.name} sshd is up"
-    }
-  end
-
-end
